[EventFilter] Move private headers in plugins dir; merge plugins cc/h files
