--- EventFilter/CSCTFRawToDigi/plugins/SealModule.cc
@@ -1,8 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include <EventFilter/CSCTFRawToDigi/interface/CSCTFUnpacker.h>
-#include <EventFilter/CSCTFRawToDigi/interface/CSCTFPacker.h>
-#include <EventFilter/CSCTFRawToDigi/interface/CSCTFAnalyzer.h>
-
-DEFINE_FWK_MODULE(CSCTFUnpacker);
-DEFINE_FWK_MODULE(CSCTFPacker);
-DEFINE_FWK_MODULE(CSCTFAnalyzer);
